+++ src/utils/fragment.py
